Some progress on a faster iterator
This method was meant to be allocation-free and not repeat any logic
but I think Julia is standing in the way. I think possibly a mutable
query iterator would be a possibility, since currently immutables
containing references to mutables are constructed on the heap,
whereas we could try implement everything in a type.